// packages/plugin-issue/src/generated-admin-types.ts
export type Maybe<T> = T;
export type InputMaybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string | number;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
  Money: any;
  Upload: any;
};
// ...
export type CampaignApplyForBeforeDataInput = {
  /** 按钮文字 */
  buttonText?: InputMaybe<Scalars['String']>;
  /** 政策小字 */
  policyText: Scalars['String'];
  /** 规则小字 */
  ruleText: Array<Scalars['String']>;
  /** 小字（标题下方） */
  subTitle: Array<Scalars['String']>;
  /** 文本框文字提示 */
  textBoxPlaceholder?: InputMaybe<Scalars['String']>;
  /** 标题（加粗文字） */
  title: Scalars['String'];
};

export type CampaignApplyForResultData = {
  __typename?: 'CampaignApplyForResultData';
  /** Expiry Date */
  expiryLabel: Scalars['String'];
  /** 规则小字 */
  ruleText: Array<Scalars['String']>;
  /** Terms条款label */
  termsCondition: Scalars['String'];
};

export type CampaignApplyForResultDataInput = {
  /** Expiry Date */
  expiryLabel: Scalars['String'];
  /** 规则小字 */
  ruleText: Array<Scalars['String']>;
  /** Terms条款label */
  termsCondition: Scalars['String'];
};
// ...
export type CampaignTranslationInput = {
  applyForBeforeData: CampaignApplyForBeforeDataInput;
  applyForResultData?: InputMaybe<CampaignApplyForResultDataInput>;
  id?: InputMaybe<Scalars['ID']>;
  languageCode: LanguageCode;
  lgImage?: InputMaybe<Scalars['String']>;
  name: Scalars['String'];
  shortDesc?: InputMaybe<Scalars['String']>;
  xsImage?: InputMaybe<Scalars['String']>;
};

export enum CampaignType {
  /** 所有用户, 固定金额拉新, 直降模式 */
  AllUserFixedAmount = 'AllUserFixedAmount',
  /** 买一送一+镜片打折 */
  BuyXGetYFreeWithLensDiscount = 'BuyXGetYFreeWithLensDiscount',
  /** 节假日直活动 */
  DirectDiscount = 'DirectDiscount',
  /** 新用户首单镜架1USD */
  NewUserFirstPair = 'NewUserFirstPair',
  /** 新用户固定金额拉新, 领券模式 */
  NewUserFixedAmount = 'NewUserFixedAmount',
}
// ...
export type CreateCampaignInput = {
  campaignType: CampaignType;
  code: Scalars['String'];
  enabled: Scalars['Boolean'];
  needClaimCoupon?: InputMaybe<Scalars['Boolean']>;
  periodOfValidity?: InputMaybe<Scalars['Int']>;
  promotionId?: InputMaybe<Scalars['Int']>;
  translations: Array<CampaignTranslationInput>;
};
// ...
export type DeletionResponse = {
  __typename?: 'DeletionResponse';
  message?: Maybe<Scalars['String']>;
  result: DeletionResult;
};

export enum DeletionResult {
  /** The entity was successfully deleted */
  DELETED = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NOT_DELETED = 'NOT_DELETED',
}
// ...
export type UpdateCampaignInput = {
  campaignType?: InputMaybe<CampaignType>;
  code?: InputMaybe<Scalars['String']>;
  enabled?: InputMaybe<Scalars['Boolean']>;
  id: Scalars['ID'];
  needClaimCoupon?: InputMaybe<Scalars['Boolean']>;
  periodOfValidity?: InputMaybe<Scalars['Int']>;
  promotionId?: InputMaybe<Scalars['Int']>;
  translations?: InputMaybe<Array<CampaignTranslationInput>>;
};